requirements.txt
aioodbc==0.4.0
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
fastapi==0.103.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
passlib==1.7.4
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
PyJWT==2.8.0
pyodbc==4.0.39
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
